Refactor OSv compiler
OSv compiler was misorganized, it was difficult to understand
the relation between osv-java.go and osv-virtualbox.go both sitting
in the same folder. Not any more - first refactored directory structure
of the pkg/compilers/osv to basically look like:
```
osv
 |- osvbootstrap
   |- virtualbox.go
   |- vmware.go
   |- aws.go
 |- java.go
```
Directly inside `osv` folder reside the actual compilers that make
use of Capstan (to compose package for selected language). Then there
is the `osvbootstrap` folder containing post-processing logic
(called "Bootstrapper") to carry out conversion of Capstan result into
appropriate image format (for the provider specified). The actual compiler
gets reference to the Bootstrapper and invokes its method
Bootstrap() to perform the conversion.

And that's it, code is more readable and managable.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>