OpenStack provider now considers logical size to pick flavor
When building unikernel with forked Capstan, user can specify
its filesystem size (10 GB by default). Actual size of the
unikernel is usually small, but can grow when the unikernel is
running (up to the filesystem size). On OpenStack we must run
pick flavor large enough to fit the filesystem size of the unikernel.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>